load mnist in float64 to match double dtype

+++ python/datasets.py
@@ -22,8 +22,8 @@ def mnist_fc(nsamples=1000):
     (train_batch, train_labels), (test_batch, test_labels) = mnist.load_data()
 
     # Reshape Data
-    train_batch = train_batch.reshape(train_batch.shape[0], 1, 28*28).astype("float32")[:nsamples]
-    test_batch  = test_batch .reshape(test_batch .shape[0], 1, 28*28).astype("float32")[:nsamples]
+    train_batch = train_batch.reshape(train_batch.shape[0], 1, 28*28).astype("float64")[:nsamples]
+    test_batch  = test_batch .reshape(test_batch .shape[0], 1, 28*28).astype("float64")[:nsamples]
 
     train_batch /= 255 # normalize pixel values
     test_batch  /= 255
